atualização 20/10/2024

+++ src/aula04_EstruturaDecisao/exercicios/exercicio04.java
@@ -5,7 +5,7 @@ import java.util.Scanner;
 public class exercicio04 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        int vendas;
+        double vendas;
         double porcentagem;
         double comissao;
